+++ src/main/java/ru/mail/polis/ya111/DAOimpl.java
@@ -1,0 +1,37 @@
+package ru.mail.polis.ya111;
+
+import org.jetbrains.annotations.NotNull;
+import ru.mail.polis.DAO;
+import ru.mail.polis.Record;
+
+import java.io.IOException;
+import java.nio.ByteBuffer;
+import java.util.Iterator;
+import java.util.SortedMap;
+import java.util.TreeMap;
+
+public class DAOimpl implements DAO {
+
+    private final SortedMap<ByteBuffer, Record> data = new TreeMap<>();
+
+    @NotNull
+    @Override
+    public Iterator<Record> iterator(@NotNull final ByteBuffer from) throws IOException {
+        return data.tailMap(from).values().iterator();
+    }
+
+    @Override
+    public void upsert(@NotNull final ByteBuffer key, @NotNull final ByteBuffer value) throws IOException {
+        data.put(key, Record.of(key, value));
+    }
+
+    @Override
+    public void remove(@NotNull final ByteBuffer key) throws IOException {
+        data.remove(key);
+    }
+
+    @Override
+    public void close() throws IOException {
+        //TODO
+    }
+}
